+++ test/unit/node-helper.js
@@ -2,7 +2,6 @@
 
 import Vips from '../../lib/vips-node.mjs';
 
-import { tmpdir } from 'node:os';
 import { expect } from 'chai';
 
 globalThis.expect = expect;
@@ -26,10 +25,6 @@ export async function mochaGlobalSetup () {
 
       // Hide warning messages
       module.ENV.VIPS_WARNING = 0;
-
-      // libvips stores temporary files by default in `/tmp`;
-      // set the TMPDIR env variable to override this directory
-      module.ENV.TMPDIR = tmpdir();
     }
   };
   globalThis.vips = await Vips(options);